+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.2.0 (2022-11-22)
+* Added Controller structure
+* Updated Genesis for Fluidity Cast
+
 ## v0.1.1 (2022-11-22)
 * Added cast customisation to Hub